Agrego servicio de Guardar Historial de Modificacion de Insumo

+++ Code/ProyectoIndustrial/Proyecto-ejb/src/main/java/Supply/Supply.java
@@ -33,7 +33,7 @@ public class Supply implements Serializable{
     @Column(name = "cost",scale = 2)
     private Double cost;
     @Column(name = "quantity",scale = 2)
-    private Double quantity;
+    private Integer quantity;
     @Column(name = "availability")
     private boolean availability;
     @Column(name = "description")
@@ -46,7 +46,7 @@ public class Supply implements Serializable{
     public Supply() {
     }
 
-    public Supply(Integer code, String name, LocalDate expirationDate, LocalDate dateOfAdmission, Double cost, Double quantity, boolean availability, String description, Measure measure) {
+    public Supply(Integer code, String name, LocalDate expirationDate, LocalDate dateOfAdmission, Double cost, Integer quantity, boolean availability, String description, Measure measure) {
         this.code = code;
         this.name = name;
         this.expirationDate = expirationDate;
@@ -98,11 +98,11 @@ public class Supply implements Serializable{
         this.cost = cost;
     }
 
-    public Double getQuantity() {
+    public Integer getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Double quantity) {
+    public void setQuantity(Integer quantity) {
         this.quantity = quantity;
     }
